test(fe): organize generated API e2e flows by screen
